requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
Django==3.2
django-debug-toolbar==3.4.0
django-split-settings==1.1.0
elasticsearch==7.7.0
Faker==14.2.0
flake8==4.0.0
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uWSGI==2.0.20
redis==4.3.4
pydantic==1.10.2
backoff==2.2.1
